Rename: 스타일 파일 이름 재설정
